No longer use preregistration and setup script. Nodekeys pulled from hiera.
Big changes today! New node installations no longer use host pre-registration
to set up things on the master server, nor are node keys built with the handshake
that goes with that mechanism. Instead, Hiera is now the repository for nodekeys
for our client systems. For the time being they are unencrypted, but we'll work
on that! We may even use the same mechanism for root passwords later on!

The best part: our Puppet runs are now fully predictable! Up to this point, you
would always need 2x2 Puppet runs for a complete client setup (generate pre-reg
on client, perform prereg on master, install and configure on client, make root
account on master). That's no longer needed! The run on the client sets up all
the needed executables for the master in one go. And the master will then run
them all in one go! Wonderful!