Root passwords now pulled from Hiera. Did some cleanup on older files.
Because we're already putting important information into Hiera, I decided
to add the root password (hash) as well.

I have cleaned up params.pp, taking out a lot of stuff that was still specific
to an organisation and other things pertaining to the host preregistration which
we're not doing any more.